vv
last v1

+++ ft_strlen.c
@@ -6,14 +6,15 @@
 /*   By: azgaoua <azgaoua@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/31 11:58:53 by azgaoua           #+#    #+#             */
-/*   Updated: 2022/10/17 18:46:25 by azgaoua          ###   ########.fr       */
+/*   Updated: 2022/11/17 15:49:59 by azgaoua          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+
 #include "libft.h"
 
 size_t	ft_strlen(const char *str)
 {
-	unsigned long long	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
